+++ src/bwStore.py
@@ -81,8 +81,7 @@ class BrikStore():
 
             if char == '\\':
                 argB.append(ps.string[ps.pos:ps.pos+2])
-                ps.pos += 2
-                continue
+                ps.pos += 1
 
             elif char == '[':
                 brikStack.append(ps.pos)
@@ -102,7 +101,8 @@ class BrikStore():
 
             ps.pos += 1
         
-        raise UnclosedBrikError(context.elem, context.prop, ps.source)
+        print('parseBrikArg')
+        raise UnclosedBrikError(context.elem, context.prop, ps.string)
 
 
     def evalBrik(self, ps) -> str:
@@ -119,11 +119,9 @@ class BrikStore():
         )
         while ps.pos < len(ps.string):
             char = ps.string[ps.pos]
-
             if char == '\\':
                 nameB.append(ps.string[ps.pos:ps.pos+2])
-                ps.pos += 2
-                continue
+                ps.pos += 1
                 
             elif char == '|':
                 ps.pos += 1
@@ -136,7 +134,7 @@ class BrikStore():
                 nameB.append(char)
 
             ps.pos += 1
-
+        print('evalBrik')
         raise UnclosedBrikError(context.elem, context.prop, ps.string)
 
     def evalValue(self, string:str) -> str:
@@ -152,8 +150,7 @@ class BrikStore():
         
             if char == '\\':
                 result.append(ps.string[ps.pos:ps.pos+2])
-                ps.pos += 2
-                continue
+                ps.pos += 1
                 #ignore escapes
         
             elif char == '[':
@@ -248,13 +245,16 @@ def underlineBrik(context, string):
 @BrikStore.addStdlib('dup', 2)
 def repeatBrik(context, times, value):
     nTimes = context.parse(times)
-    times = asNum(times, err=InvalidArgError(
-        context.elem, context.prop, 'dup', 'TIMES', nTimes
-    ))
+    unit = Unit.fromStr(nTimes, signs='+0', units=('',))
+    if unit is None:
+        raise InvalidArgError(context.elem, context.prop, 'dup', 'TIMES', nTimes)
+    #times = asNum(times, err=)
+    times = unit.toInt()
+    ofs = 0 if unit.sign == '0' else 1
     result = []
     newStore = context.store.copy()
     for i in range(times):
-        newStore.add('d', str(i+1))#FIXME
+        newStore.add('d', str(i+ofs))
         result.append(newStore.parse(value))
     return ''.join(result)
     
@@ -280,16 +280,49 @@ def upperBrik(context, value):
         return m.group(1).lower()
     return re.sub(r'(?<!\\)([A-Z])', repl, value)
 
-@BrikStore.addStdlib('substring', 3)
+@BrikStore.addStdlib('substr', 3)
 def subBrik(context, value, start, length):
     value = context.parse(value)
-    start = asNum(context.parse(start), err=InvalidArgError(
-        context.elem, context.prop, 'substring', 'START', start
-    ))-1
-    length = asNum(context.parse(length), err=InvalidArgError(
-        context.elem, context.prop, 'substring', 'LENGTH', length
-    ))
+    
+    start = context.parse(start)
+    startUnit = Unit.fromStr(context.parse(start), signs='+0', units=('',))
+    if startUnit is None:
+        raise InvalidArgError(context.elem, context.prop, 'substr', 'START', start)
+    start = startUnit.toInt()
+    if startUnit.sign != '0':
+        start -= 1
+    
+    length = context.parse(length)
+    lengthUnit = Unit.fromStr(context.parse(length), signs='+0', units=('',))
+    if lengthUnit is None:
+        raise InvalidArgError(context.elem, context.prop, 'substr', 'LENGTH', length)
+    length = lengthUnit.toInt()
+    
     return value[start:start+length]
+
+@BrikStore.addStdlib('slice', (2, 3))
+def sliceBrik(context, value, start, stop=None):
+    value = context.parse(value)
+
+    start = context.parse(start)
+    startUnit = Unit.fromStr(context.parse(start), signs='+-0', units=('',))
+    if startUnit is None:
+        raise InvalidArgError(context.elem, context.prop, 'slice', 'START', start)
+    start = startUnit.toInt()
+    if startUnit.sign in ('+', ''):
+        start -= 1
+
+    if stop is not None:
+        stop = context.parse(stop)
+        stopUnit = Unit.fromStr(context.parse(stop), signs='+-0', units=('',))
+        if stopUnit is None:
+            raise InvalidArgError(context.elem, context.prop, 'slice', 'STOP', stop)
+        stop = stopUnit.toInt()
+        if stopUnit.sign in ('+', ''):
+            stop -= 1
+        
+    return value[start:stop]
+
 
 @BrikStore.addStdlib('/', 1)
 def expansionMacro(context, value):
@@ -306,12 +339,14 @@ def comparisonMacro(context, value):
         raise bWError("'{value}' does not contain a valid comparison",
         prop=context.prop, elem=context.elem, value=value
         )
-    left = asNum(left.strip(), err=InvalidArgError(
-        context.elem, context.prop, '?', 'OPERAND', left
-    ))
-    right = asNum(right.strip(), err=InvalidArgError(
-        context.elem, context.prop, '?', 'OPERAND', right
-    ))
+    leftUnit = Unit.fromStr(left.strip(), units='all')
+    if leftUnit is None:
+        raise InvalidArgError(context.elem, context.prop, '?', 'OPERAND', left)
+    rightUnit = Unit.fromStr(right.strip(), units='all')
+    if rightUnit is None:
+        raise InvalidArgError(context.elem, context.prop, '?', 'OPERAND', right)
+    left = leftUnit.toFloat()
+    right = rightUnit.toFloat()
     op = op.strip()
     if op == '==':
         final = left == right
@@ -363,7 +398,8 @@ def mathMacro(context, value):
 
     for token in tokens:
         if token == '': continue
-        currentToken = asNum(token)
+        currentToken = Unit.fromStr(token, units='all')
+        #currentToken = asNum(token)
         if currentToken is None:
             if token in '*/%':
                 op = token
@@ -379,6 +415,7 @@ def mathMacro(context, value):
             else:
                 raise InvalidArgError(context.elem, context.prop, '#', 'OPERATION',token)
         else:
+            currentToken = currentToken.toFloat()
             if op is None:
                 accum = currentToken
             elif accum is None:
@@ -398,8 +435,6 @@ def mathMacro(context, value):
     op = None
 
     for token in newTokens:
-        #currentToken = asNum(token)
-        #if currentToken is None:
         if type(token) not in (int, float):
             op = token
         else: