+++ src/bwUtils.py
@@ -1,4 +1,5 @@
 
+from dataclasses import dataclass
 import re
 from collections.abc import Mapping
 from collections import UserDict
@@ -24,18 +25,10 @@ class bWError(Exception):
             self.msg = "error from unknown source:\n\t"+msg
         self.kwargs = kwargs
 
-
     @property
     def message(self):
         return self.msg.format(**self.kwargs)
 
-# I actually want to make this not even an error, it just gets ignored
-class InvalidPropError(bWError):
-    def __init__(self, elem, prop):
-        super().__init__("'{prop}' is not a valid property for this element", 
-        elem=elem, prop=prop
-        )
-    
 class InvalidValueError(bWError):
     def __init__(self, elem, prop, value):
         super().__init__("'{value}' is not a valid value for this property",
@@ -46,6 +39,12 @@ class InvalidArgError(bWError):
     def __init__(self, elem, prop, brik, arg, value):
         super().__init__("'{value}' is not a valid {arg} argument for brik [{brik}| ]",
         elem=elem, prop=prop, brik=brik, arg=arg, value=value
+        )
+
+class UnclosedBrikError(bWError):
+    def __init__(self, elem, prop, source):
+        super().__init__("'{source}' has an unclosed brik",
+        elem=elem, prop=prop, source=source
         )
         
 class bWSyntaxError(bWError):
@@ -69,12 +68,6 @@ class UnexpectedEOFError(bWSyntaxError):
     def __init__(self, file, name):
         super.__init__("unexpected EOF in '{name}'",
         file=file, name=name)
-
-class UnclosedBrikError(bWSyntaxError):
-    def __init__(self, elem, prop, source):
-        super().__init__("'{source}' has an unclosed brik",
-        elem=elem, prop=prop, source=source
-        )
 
 class Collection(SimpleNamespace):
     '''Used to hold generated elements'''
@@ -102,20 +95,86 @@ class AttrDict():
         return copy
 
 
-def asNum(string:str, *, err:bWError = False) -> Union[int, Literal[None]]:
-    '''tries to convert a number string to an int
-    if a number isn't found raise an error if present else return None'''
-    if re.match(r'^-?\d+$', string):
-        return int(string)
-    elif re.match(r'^\d*(\.\d+)? *in$', string):
-        #FIXME this uses a hardcoded dpi
-        #I'll need to get context in here somehow lol
-        return int(float(string[:-2].strip())*300)
-    else:
-        if err:
-            raise err
+
+@dataclass
+class Unit():
+    '''Unit represents a unitized number'''
+    sign: str
+    num: int
+    unit: str
+    def toFloat(self, **data):
+        'return the number as an int according to type'
+        if self.unit in ('px', 'pt', ''):
+            return self.num
+        elif self.unit == 'in':
+            if 'dpi' in data:
+                return self.num*data['dpi']
+            else:
+                return self.num
+        elif self.unit == 'mm':
+            if 'dpi' in data:
+                return (self.num/25.4)*data['dpi']
+            else:
+                return self.num
+        elif self.unit == '%':
+            if 'whole' in data:
+                return self.num/100*data['whole']
+            else:
+                return self.num
+    def toInt(self, **data):
+        return int(self.toFloat(**data))
+    
+    @staticmethod
+    def fromStr(string, signs='-+', units=('px', 'in', 'mm')):
+        if units == 'all':
+            units = ('px', 'pt', 'in', 'mm', '%')
+
+        signs = r'^(?P<sign>['+signs+r']?)'
+        unitRe = r' *(?P<unit>(?:' + '|'.join(units) + r')?)$'
+        whole = signs + r'(?P<num>\d+)' + unitRe
+        flt = signs + r'(?P<num>\d*\.\d+)' + unitRe
+        frac = signs + r'(?P<whole>\d+) (?P<numer>\d+)/(?P<denom>\d+)' + unitRe
+        fracOnly = signs + r'(?P<numer>\d+)/(?P<denom>\d+)' + unitRe
+
+        if (match := re.match(whole, string)) or (match := re.match(flt, string)):
+            sign = match.group('sign')
+            num = float(match.group('num'))
+            unit = match.group('unit')
+        elif match := re.match(fracOnly, string):
+            sign = match.group('sign')
+            numer = match.group('numer')
+            denom = match.group('denom')
+            unit = match.group('unit')
+            num = (int(numer)/int(denom))
+        elif match := re.match(frac, string):
+            sign = match.group('sign')
+            whole = match.group('whole')
+            numer = match.group('numer')
+            denom = match.group('denom')
+            unit = match.group('unit')
+            num = int(whole)+(int(numer)/int(denom))
         else:
             return None
+        
+        if unit == '':
+            unit = units[0]
+        if sign == '-' and unit != '%':
+            num *= -1
+        return Unit(sign, num, unit)
+
+
+#def asNum(string:str, *, err:bWError = False) -> Union[int, Literal[None]]:
+#    '''tries to convert a number string to an int
+#    if a number isn't found raise an error if present else return None'''
+#    if re.match(r'^-?\d+$', string):
+#        return int(string)
+#    elif re.match(r'^\d*(\.\d+)? *in$', string):
+#        #FIXME this uses a hardcoded dpi
+#        #I'll need to get context in here somehow lol
+#        return int(float(string[:-2].strip())*300)
+#    else:
+#        if err:
+#            raise err
 
 trues = 'yes on true'.split()
 falses = 'no off false 0'.split()
@@ -138,13 +197,24 @@ expansions = {
     '\\': '\\', 'n': '\n',
     't': '\t', 's': ' ',
 }
+
 def evalEscapes(string:str) -> str:
-    '''replace escapes with their corresponding values'''
-    def repl(m):
-        c = m.group(1)
-        return expansions.get(c, c)
-        #return self.expansions[m.group(0)]
-    return re.sub(r'\\(.)', repl, string)
+    pos = 0
+    accum = []
+    while pos < len(string):
+        char = string[pos]
+        if char == '\\':
+            pos += 1
+            char = string[pos]
+            if char in expansions:
+                accum.append(expansions[char])
+            else:
+                accum.append(char)
+        else:
+            accum.append(char)
+        pos += 1
+    
+    return build(accum)
 
 def deepUpdate(self:Mapping, other:Mapping):
     '''like update, but if a given index is a mapping in both self and other we recurse'''
@@ -209,11 +279,7 @@ class CSVParser():
         while pos < len(line):
             char = line[pos]
 
-            if char == '\\':
-                accum.append(self.processEscape(line,pos))
-                pos += 1
-            
-            elif char == ',':
+            if char == ',':
                 name = build(accum)
                 if name != '':
                     headers.append(name)
@@ -298,7 +364,7 @@ class LayoutParser():
                 value = build(accum)
                 return value
             
-            if char == '}':
+            elif char == '}':
                 value = build(accum)
                 self.pos -= 1 #OH NO A BCKTRACK
                 return value
@@ -324,11 +390,7 @@ class LayoutParser():
         while self.pos < len(self.string):
             char = self.string[self.pos]
 
-            if char == '\\':
-                accum.append(self.string[self.pos:self.pos+2])
-                self.pos += 1
-            
-            elif char == ':':
+            if char == ':':
                 name = build(accum)
                 accum = []
                 self.pos += 1
@@ -345,6 +407,11 @@ class LayoutParser():
                 if name != '':
                     raise NoValueError(self.filename, elem, name)
                 return section
+            
+            elif char == '=':
+                name = build(accum)
+                raise bWSyntaxError("'{elem}' section can not define briks",
+                elem=elem, file=self.filename)
         
             else:
                 accum.append(char)
@@ -361,11 +428,7 @@ class LayoutParser():
         while self.pos < len(self.string):
             char = self.string[self.pos]
 
-            if char == '\\':
-                accum.append(self.string[self.pos:self.pos+2])
-                self.pos += 1
-            
-            elif char == ':':
+            if char == ':':
                 name = build(accum)
                 accum = []
                 self.pos += 1
@@ -380,6 +443,11 @@ class LayoutParser():
                 if name != '':
                     raise NoValueError(self.filename, elem, name)
                 return section
+            
+            elif char == '=':
+                name = build(accum)
+                raise bWSyntaxError("'{elem}' section can not define briks",
+                elem=elem, file=self.filename)
             
             else:
                 accum.append(char)
@@ -398,11 +466,7 @@ class LayoutParser():
         while self.pos < len(self.string):
             char = self.string[self.pos]
 
-            if char == '\\':
-                accum.append(self.string[self.pos:self.pos+2])
-                self.pos += 1
-            
-            elif char == '=':
+            if char == '=':
                 name = build(accum)
                 accum = []
                 self.pos += 1
@@ -460,12 +524,8 @@ class LayoutParser():
 
         while self.pos < len(self.string):
             char = self.string[self.pos]
-
-            if char == '\\':
-                accum.append(self.string[self.pos:self.pos+2])
-                self.pos += 1
             
-            elif char == '{':
+            if char == '{':
                 name = build(accum)
                 self.pos += 1
                 accum = []
@@ -499,9 +559,41 @@ class LayoutParser():
 
 
 if __name__ == '__main__':
-    ad = AttrDict()
-    ad.foo = 5
-    print(ad['foo'])
-    ad['bar'] = 6
-    print(ad.bar)
-    print(ad)
+
+    expansions = {
+        '\\': '\\', 'n': '\n',
+        't': '\t', 's': ' ',
+    }
+
+    def eval(string:str) -> str:
+        pos = 0
+        accum = []
+        while pos < len(string):
+            char = string[pos]
+            if char == '\\':
+                pos += 1
+                char = string[pos]
+                if char in expansions:
+                    accum.append(expansions[char])
+                else:
+                    accum.append(char)
+            else:
+                accum.append(char)
+            print(build(accum))
+            pos += 1
+    
+    def parse(string:str):
+        pos = 0
+        accum = []
+        while pos < len(string):
+            char = string[pos]
+            
+            if char == '\\':
+                accum.append(string[pos:pos+2])
+                pos += 1
+            else:
+                accum.append(char)
+            print(pos, accum)
+            pos += 1
+        return build(accum)
+    eval(parse(r'te\k\kst'))