+++ docs/changelog.md
@@ -1,5 +1,11 @@
-# Change log
+# Changelog
 This page collects the changes made to brikWork over time
+
+## v 0.4
+ - added: more units! checkout [values] for how they work
+ - added: strip brik, `[$| ]`, to remove an explicit unit. This won't make sense until you read the new docs
+ - added: `[slice| STRING | START | STOP]` takes substring of `STRING` like `[substr| ]` but uses start and stop positions instead of a start position and a length
+ - changed: the `[substring| ]` brik has been renamed to `[substr| ]`
 
 ## v 0.3
  - added: A graphical interface